backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic_settings
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
pytest
httpx